# Improve validate performance - up to 95x for static rules