@cyppe 2025-04-21
Some relevant discussions here: https://github.com/spatie/laravel-data/discussions/875