@aa-dragosmocrii 2025-05-20
@rubenvanassche I would appreciate if this PR could get some attention as a priority. I have performed some tests on a real dataset and the time went down from `3m12s` to `13s700ms` by using this patch. The author also makes a good point about "static" validation rules added as attributes, and the "rules" method which allows for dynamic validation rules. In most cases, the static validation rules are more than enough, which means that this PR will bring a huge performance improvement to the majority of use cases.

@cyppe great work!